requirements_Version4.txt
python-telegram-bot==20.5
python-dotenv==1.0.0
aiohttp==3.8.4